Fluxome

Documentation for Fluxome.

Fluxome.ReactionNetworkSimulation

Fluxome.plot_and_save_distribution_timepoint

Fluxome.plot_and_save_distribution_timepoints

Fluxome.save_reaction_network_graph

Fluxome.setup_reaction_network_reports

Fluxome.simulate_reaction_network

Fluxome.ReactionNetworkSimulation — Type

ReactionNetworkSimulation(rn::ReactionSystem, reportpath::String, u0::Vector{Float64}, tspan::Tuple{Float64, Float64})

A composite type representing a reaction network, its initial state, time span, and report path for storing results.

Fields

rn::ReactionSystem: The reaction network defined as a Catalyst.jl ReactionSystem.

reportpath::String: The path where simulation results will be saved.

u0::Vector{Float64}: The initial state of the system.

tspan::Tuple{Float64, Float64}: The time span for the simulation.

Examples

using Fluxome

# Define the reaction network and the report path
rn = @reaction_network begin
	σ, 0 --> A
	d, A --> 0
end
reportpath = "reports/immigrationdeath/";

# Initialize the ReactionNetworkSimulation
u0 = zeros(50);
u0[1] = 1.0;
tspan = (0, 10.0);
network = ReactionNetworkSimulation(rn, reportpath, u0, tspan);
setup_reaction_network_reports(network);

# Save the graph
save_reaction_network_graph(network, "network_petri_graph.pdf")

# Simulate the network
ps = [10.0, 1.0];
sol = simulate_reaction_network(network, ps);

# Plot and save the distributions at different time points
plot_and_save_distribution_timepoint(sol, network, 3, "0_001")
plot_and_save_distribution_timepoint(sol, network, 12, "0_3")

# output
Python: None

Fluxome.plot_and_save_distribution_timepoint — Method

plotandsavedistributiontimepoint(sol, rn::ReactionNetwork, tidx::Int, filename_suffix::String)

Plot and save the distribution at a specific time index tidx for the given solution sol and reaction network rn. The plot is saved with a filename suffix filename_suffix.

Arguments sol: The solution of the simulation. rn::ReactionNetwork: The reaction network. tidx::Int: The time index for the distribution to plot. filename_suffix::String: The filename suffix for the saved plot.

Fluxome.plot_and_save_distribution_timepoints — Method

plotandsavedistributiontimepoints(sol, rn::ReactionNetworkSimulation, plottimes, filenamesuffix::String)

Plot and save the distributions at specific times for the given solution sol and reaction network rn. The times at which the distributions are to be plotted are specified in plot_times. The plot is saved with a filename suffix filename_suffix.

Arguments

sol: The solution of the simulation.

rn::ReactionNetworkSimulation: The reaction network simulation.

plot_times::Vector{Float64}: A vector of times at which the distributions are to be plotted.

filename_suffix::String: The filename suffix for the saved plot.

Example

plot_and_save_distribution_timepoints(sol, rn, [1.0, 2.0, 3.0], "suffix")

This will create a plot with subplots showing the distributions at times 1.0, 2.0, and 3.0 and save it with a filename that includes the specified suffix.

Fluxome.save_reaction_network_graph — Method

savereactionnetwork_graph(rn::ReactionNetwork, filename::String)

Save the graph representation of the given reaction network rn with the specified filename filename.

Fluxome.setup_reaction_network_reports — Method

setup_reaction_network_reports(rn::ReactionNetworkSimulation)

Create the report directory for a given ReactionNetworkSimulation instance.

This function creates a directory at the specified reportpath of the ReactionNetworkSimulation instance if it does not already exist.

Arguments

rn::ReactionNetworkSimulation: The ReactionNetworkSimulation instance for which the report directory needs to be created.

Fluxome.simulate_reaction_network — Method

simulatereactionnetwork(rn::ReactionNetworkSimulation, p::Vector{Float64})

Simulate a given reaction network rn with parameters p.

Arguments

rn::ReactionNetworkSimulation: The reaction network to simulate. p::Vector{Float64}: The parameters of the reaction network.

Returns

sol: The solution of the simulation.